refactoring: project format
