+++ repository/locations/locations.go
@@ -2,58 +2,24 @@ package locations
 
 import (
 	"context"
-	"github.com/YusufOzmen01/veri-kontrol-backend/core/sources"
-	"github.com/YusufOzmen01/veri-kontrol-backend/repository/users"
+	"github.com/acikkaynak/veri-toplama-backend/db/sources"
 	"github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
 )
-
-type Repository interface {
-	GetLocations(ctx context.Context) ([]*LocationDB, error)
-	ResolveLocation(ctx context.Context, location *LocationDB) error
-	IsResolved(ctx context.Context, locationID int) (bool, error)
-	IsDuplicate(ctx context.Context, tweetContents string) (bool, error)
-	GetDocumentsWithNoTweetContents(ctx context.Context) ([]*LocationDB, error)
-}
-
-type repository struct {
-	mongo sources.MongoClient
-}
-
-func NewRepository(mongo sources.MongoClient) Repository {
-	return &repository{
-		mongo: mongo,
-	}
-}
-
-type Location struct {
-	EntryID          int       `json:"entry_id"`
-	Loc              []float64 `json:"loc"`
-	Epoch            int       `json:"epoch"`
-	OriginalMessage  string    `json:"original_message"`
-	OriginalLocation string    `json:"original_location"`
-}
 
 const (
 	TypeWreckage   = 1
 	TypeSupplyHelp = 2
 )
 
-type LocationDB struct {
-	ID               primitive.ObjectID `json:"_id" bson:"_id"`
-	EntryID          int                `json:"entry_id" bson:"entry_id"`
-	Sender           *users.User        `json:"sender" bson:"sender"`
-	Location         []float64          `json:"location" bson:"location"`
-	Corrected        bool               `json:"corrected" bson:"corrected"`
-	Verified         bool               `json:"verified" bson:"verified"`
-	OriginalAddress  string             `json:"original_address" bson:"original_address"`
-	CorrectedAddress string             `json:"corrected_address" bson:"corrected_address"`
-	OpenAddress      string             `json:"open_address" bson:"open_address"`
-	Apartment        string             `json:"apartment" bson:"apartment"`
-	Type             int                `json:"type" bson:"type"`
-	Reason           string             `json:"reason" bson:"reason"`
-	TweetContents    string             `json:"tweet_contents" bson:"tweet_contents"`
+type repository struct {
+	mongo sources.MongoClient
+}
+
+func NewRepository(mongo sources.MongoClient) *repository {
+	return &repository{
+		mongo: mongo,
+	}
 }
 
 func (r *repository) GetLocations(ctx context.Context) ([]*LocationDB, error) {
