+++ services/feeds.go
@@ -2,51 +2,41 @@ package services
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	locationsRepository "github.com/YusufOzmen01/veri-kontrol-backend/repository/locations"
-	"github.com/YusufOzmen01/veri-kontrol-backend/util"
+	locationsRepository "github.com/acikkaynak/veri-toplama-backend/repository/locations"
+	"github.com/acikkaynak/veri-toplama-backend/util"
 	"github.com/gofiber/fiber/v2"
 	"github.com/sirupsen/logrus"
-	"io"
 	"math/rand"
-	"net/http"
-	"time"
 )
 
 const afetharitaURL = "https://apigo.afetharita.com/feeds/areas?ne_lat=39.91618777305531&ne_lng=47.85149904303703&sw_lat=36.07272886939253&sw_lng=23.872389299415502"
-
-type httpClient interface {
-	Do(req *http.Request) (*http.Response, error)
-}
-
-type cache interface {
-	Get(key interface{}) (interface{}, bool)
-	Set(key, value interface{}, cost int64) bool
-	SetWithTTL(key, value interface{}, cost int64, ttl time.Duration) bool
-}
 
 type repository interface {
 	IsDuplicate(ctx context.Context, tweetContents string) (bool, error)
 }
 
-type feedServices struct {
-	repository repository
-	httpClient httpClient
-	cache      cache
+type locationServiceClient interface {
+	GetLocation(locationID int) (*singleResponse, error)
+	GetLocations(c *fiber.Ctx) ([]*locationsRepository.Location, error)
 }
 
-func NewFeedsServices(repository repository, httpClient httpClient, cache cache) *feedServices {
-	return &feedServices{
-		repository: repository,
-		httpClient: httpClient,
-		cache:      cache,
+type feedService struct {
+	repository            repository
+	locationServiceClient locationServiceClient
+	httpClient            httpClient
+}
+
+func NewFeedsServices(repository repository, locationServiceClient locationServiceClient) *feedService {
+	return &feedService{
+		repository:            repository,
+		locationServiceClient: locationServiceClient,
 	}
 }
 
-func (f *feedServices) GeetFeeds(c *fiber.Ctx) (*FeetFeedsResponse, error) {
+func (f *feedService) GeetFeeds(c *fiber.Ctx) (*FeetFeedsResponse, error) {
 	logrus.Infoln("Pulling entries")
-	locs, err := f.GetAllLocations(c)
+	locs, err := f.locationServiceClient.GetLocations(c)
 	if err != nil {
 		logrus.Errorf("Couldn't get all locations: %s", err)
 		return nil, fmt.Errorf("Couldn't get all locations: %s", err)
@@ -72,80 +62,7 @@ func (f *feedServices) GeetFeeds(c *fiber.Ctx) (*FeetFeedsResponse, error) {
 	return res, nil
 }
 
-func (f *feedServices) GetAllLocations(c *fiber.Ctx) ([]*locationsRepository.Location, error) {
-	var d struct {
-		Locations []*locationsRepository.Location `json:"results"`
-	}
-
-	data, exists := f.cache.Get("locations")
-	if exists {
-		return data.([]*locationsRepository.Location), nil
-	}
-
-	req, err := http.NewRequest("GET", afetharitaURL, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36")
-
-	resp, err := f.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := json.Unmarshal(body, &d); err != nil {
-		return nil, err
-	}
-
-	f.cache.SetWithTTL("locations", d.Locations, int64(time.Minute*15), 0)
-
-	return d.Locations, nil
-}
-
-func (f *feedServices) getLocation(locationID int) (*singleResponse, error) {
-	data, exists := f.cache.Get(fmt.Sprintf("single_location_%d", locationID))
-	if exists {
-		return data.(*singleResponse), nil
-	}
-
-	url := fmt.Sprintf("https://apigo.afetharita.com/feeds/%d", locationID)
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36")
-
-	resp, err := f.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	singleData := singleResponse{}
-	if err := json.Unmarshal(body, &singleData); err != nil {
-		return nil, err
-	}
-
-	f.cache.Set(fmt.Sprintf("single_location_%d", locationID), singleData, 0)
-
-	return &singleData, nil
-}
-
-func (f *feedServices) filterLocations(c *fiber.Ctx, locations []*locationsRepository.Location) ([]*locationsRepository.Location, error) {
+func (f *feedService) filterLocations(c *fiber.Ctx, locations []*locationsRepository.Location) ([]*locationsRepository.Location, error) {
 	locs := f.filterLocationsByEntryID(locations)
 	if len(locs) == 0 {
 		return nil, fmt.Errorf("locations is empty")
@@ -165,7 +82,7 @@ func (f *feedServices) filterLocations(c *fiber.Ctx, locations []*locationsRepos
 	return filteredLocations, nil
 }
 
-func (f *feedServices) filterLocationsByEntryID(locations []*locationsRepository.Location) []*locationsRepository.Location {
+func (f *feedService) filterLocationsByEntryID(locations []*locationsRepository.Location) []*locationsRepository.Location {
 	processedIDs := make([]int, 0)
 
 	for _, loc := range locations {
@@ -185,7 +102,7 @@ func (f *feedServices) filterLocationsByEntryID(locations []*locationsRepository
 	return locations
 }
 
-func (f feedServices) filterLocationsByCityID(c *fiber.Ctx, locations []*locationsRepository.Location) ([]*locationsRepository.Location, error) {
+func (f feedService) filterLocationsByCityID(c *fiber.Ctx, locations []*locationsRepository.Location) ([]*locationsRepository.Location, error) {
 	cityID := c.QueryInt("city_id")
 	if cityID < 1 {
 		return nil, fmt.Errorf("city id is not found")
@@ -203,7 +120,7 @@ func (f feedServices) filterLocationsByCityID(c *fiber.Ctx, locations []*locatio
 	return filteredLocations, nil
 }
 
-func (f feedServices) filterLocationsByStartingAt(c *fiber.Ctx, locations []*locationsRepository.Location) ([]*locationsRepository.Location, error) {
+func (f feedService) filterLocationsByStartingAt(c *fiber.Ctx, locations []*locationsRepository.Location) ([]*locationsRepository.Location, error) {
 	startingAt := c.QueryInt("starting_at")
 	if startingAt < 1 {
 		return nil, fmt.Errorf("startingAt is not found")
@@ -220,7 +137,7 @@ func (f feedServices) filterLocationsByStartingAt(c *fiber.Ctx, locations []*loc
 	return filteredLocations, nil
 }
 
-func (f *feedServices) getSelectedLocation(locations []*locationsRepository.Location) *locationsRepository.Location {
+func (f *feedService) getSelectedLocation(locations []*locationsRepository.Location) *locationsRepository.Location {
 	var (
 		selected  *locationsRepository.Location
 		fullText  string
@@ -240,7 +157,7 @@ func (f *feedServices) getSelectedLocation(locations []*locationsRepository.Loca
 
 		s := locations[randIndex]
 
-		singleData, err := f.getLocation(s.EntryID)
+		singleData, err := f.locationServiceClient.GetLocation(s.EntryID)
 		if err != nil {
 			logrus.Errorln(err)
 			break
@@ -270,9 +187,4 @@ func (f *feedServices) getSelectedLocation(locations []*locationsRepository.Loca
 type FeetFeedsResponse struct {
 	Count    int                           `json:"count"`
 	Location *locationsRepository.Location `json:"location"`
-}
-
-type singleResponse struct {
-	FullText         string `json:"full_text"`
-	FormattedAddress string `json:"formatted_address"`
 }
